// planning-idsdp-uaa-ui/src/app/main/modules/auth/features/model/show-marks.ts
export class ShowMarks {

    graduateResultMark: number;
    isGovEmployeeMark: number;
    postGraduateResultMark: number;
    isThesisGroupMark: number;
    researchExperienceMark: number;
    journalPublicationLocQtyMark: number;
    journalPublicationIntQtyMark: number;
    mPhilMark: number;
    PHDMark: number;
    ageMark: number;
    age: number;




    id: number;
    uuid: string;

    personName: string;
    stResearchCatTypeId: number;
    dateOfBirth: Date;
    graduateResult: string;
    isGovEmployee: boolean;
    postGraduateResult: string;
    isThesisGroup: boolean;
    isMPhil: boolean;
    isPhD: boolean;
    totalResearchExperience: number;
    journalPublicationLocQty: number;
    journalPublicationIntQty: number;
    isResearchTraining: boolean;
    nidNumber: string;
    nidVerificationStatus: string;

}